fix: moved traning status class into its own code file
