+++ cypress/support/components/ModelPage.js
@@ -26,146 +26,6 @@ export function NavigateToMyModels() { cy.Get('[data-testid="app-index-nav-link-
 export function VerifyHomeLinkShowsIncidentTriangle() { cy.Get('[data-testid="app-index-nav-link-home"]').find('i[data-icon-name="IncidentTriangle"]') }
 export function VerifyHomeLinkDoesNotShowIncidentTriangle() { cy.Get('[data-testid="app-index-nav-link-home"]').DoesNotContain('i[data-icon-name="IncidentTriangle"]') }
 
-// To validate that this code works, search src\actions\appActions.ts for these and alter them:
-//   fetchApplicationTrainingStatusThunkAsync
-//   interval:
-//   maxDuration:
-export class TrainingStatus {
-// PUBLIC:  
-  constructor() {
-    this.canRefreshTrainingStatusTime = 0
-    this.trainingStatusElements = []
-    this.trainingStatusHtml = ''
-    this.queued = false
-    this.running = false
-    this.pollingStoppedWarning = false
-    this.failed = false
-    this.completed = false
-    this.currentStatus = "unknown"
-    
-    this._GetTrainingStatus()
-
-    if (this.completed) {
-      // This is most likely due to the previous training. So we need to wait until we see a different
-      // status, most likely queued or running, before we can accept a completed status, however
-      // 4 seconds is the longest we will wait for that before we call it complete.
-      this.waitForTrainingStatusQueuedOrRunningTime = new Date().getTime() + 4000
-    } else {
-      this.waitForTrainingStatusQueuedOrRunningTime = 0
-    }
-  }
-  
-  WaitForCompleted() { this._RetryWaitForRunning() }
-
-// PRIVATE:
-  // Wait retry loop for any state other than running. 
-  // This loop can wait as long as 5 minutes.
-  _RetryWaitForRunning() {
-    cy.log('Training Status is NOT Running - Waiting for it Start Running')
-    cy.wrap(1, { timeout: 5 * 60 * 1000 }).should(() => {
-      if (this._MonitorTrainingStatus()) { 
-        return // Because the status is now complete!
-      }
-
-      if (!this.running) { 
-        throw new Error(`Stauts is ${this.currentStatus} - Still Waiting for Status == Running or Completed`) 
-      }
-
-      // The status is now Running so we need to wait a different 
-      // amount of time for it to complete running.
-      this._RetryWaitForCompleted()
-    })
-  }
-
-  // Wait retry loop while in the running state. 
-  // This loop can wait as long as 40 seconds.
-  _RetryWaitForCompleted() {
-    cy.log('Training Status is Running - Waiting for it to Complete')
-    cy.wrap(1, { timeout: 40 * 1000 }).should(() => {
-      if (this._MonitorTrainingStatus()) { 
-        return // Because the status is now complete!
-      }
-      
-      if (this.running) { 
-        throw new Error(`Stauts is ${this.currentStatus} - Still Waiting for Status == Completed`)
-      }
-
-      // The status is no longer Running so we need to wait a different 
-      // amount of time for it to start running again.
-      this._RetryWaitForRunning()
-    })
-  }
-
-  _GetTrainingStatus() {
-    const funcName = '_GetTrainingStatus'
-    const priorTrainingStatusElementsLength = this.trainingStatusElements.length
-    const priorTraingStatusHtml = this.trainingStatusHtml
-
-    // Get the Training Status Elements from the DOM
-    this.trainingStatusElements = Cypress.$('[data-testid^="training-status-"]')
-    
-    // Turn the elements into a text string that can be searched and dumped to the log.
-    this.trainingStatusHtml = ''
-    for (let i = 0; i < this.trainingStatusElements.length; i++) {
-      this.trainingStatusHtml += '\n' + this.trainingStatusElements[i].outerHTML
-    }
-
-    // Determine if anything changed.
-    let changed = this.trainingStatusElements.length != priorTrainingStatusElementsLength || 
-                  this.trainingStatusHtml.length != priorTraingStatusHtml.length || 
-                  this.trainingStatusHtml != priorTraingStatusHtml
-
-    if (changed) {
-      helpers.ConLog(funcName, `Number of Training Status Elements Found: ${this.trainingStatusElements.length} - Elements: ${this.trainingStatusHtml}`)
-      
-      this.queued = this._TrainingStatusContains('queued')
-      this.running = this._TrainingStatusContains('running')
-      this.pollingStoppedWarning = this._TrainingStatusContains('polling-stopped-warning')
-      this.failed = this._TrainingStatusContains('failed')
-      this.completed = this._TrainingStatusContains('completed')
-    }
-  }
-
-  _TrainingStatusContains(target) { 
-    const returnValue = this.trainingStatusHtml.includes(`data-testid="training-status-${target}`)
-    helpers.ConLog('_TrainingStatusContains', `'${target}': ${returnValue}`)
-    if (returnValue) { this.currentStatus = target }
-    return returnValue
-  }
-
-  _MonitorTrainingStatus() {
-    const funcName = 'TrainingStatus._MonitorTrainingStatus'
-    this._GetTrainingStatus()
-    const currentTime = new Date().getTime()
-
-    if (this.completed) {
-      if (currentTime > this.waitForTrainingStatusQueuedOrRunningTime) {
-        helpers.ConLog(funcName, 'Training Status IS COMPLETED!')
-        return true
-      }
-      helpers.ConLog(funcName, 'Waiting to see queued or running before we can accept the current completed status')
-      return false
-    }
-    this.waitForTrainingStatusQueuedOrRunningTime = 0
-
-    if ((this.pollingStoppedWarning || this.failed) && (currentTime > this.canRefreshTrainingStatusTime)) {
-      // Setting this allows the UI to do some work and change the status before we click the 
-      // refresh button again too soon.
-      this.canRefreshTrainingStatusTime = currentTime + 2000
-
-      helpers.ConLog(funcName, 'Click the Refresh Button...')
-
-      // When we get here it is possible there are two refresh buttons on the page, one that
-      // is covered up by a popup dialog, so we need to click on the last one found.
-      const elements = Cypress.$('button[data-testid="training-status-refresh-button"]')
-      Cypress.$(elements[elements.length - 1]).click()
-    }
-    
-    return false
-  }
-}
-
-
 // Verify for just the Left Pane "Train Dialogs" link.
 export function VerifyNoErrorTriangleOnPage() { VerifyIncidentIcon(false, 'cl-color-error') }
 export function VerifyErrorTriangleForTrainDialogs() { VerifyIncidentIcon(true, 'cl-color-error') }
